refactor: Move final Python files from root and update imports
- Moved superadmin_api.py and sync_api.py to src/admin/
- Moved client_mcp.py to examples/
- Moved validate_format_models.py to scripts/dev/
- Updated all imports for moved API files
- Root directory now much cleaner with only 4 Python files moved
