+++ tests/integration/test_superadmin_api_integration.py
@@ -5,7 +5,7 @@
 import pytest
 from flask import Flask
 
-from superadmin_api import superadmin_api
+from src.admin.superadmin_api import superadmin_api
 
 pytestmark = pytest.mark.integration
